+++ src/compas_timber/connections/__init__.py
@@ -2,22 +2,28 @@ from .french_ridge_lap import FrenchRidgeLapJoint
 from .joint import BeamJoinningError
 from .joint import Joint
 from .joint import beam_side_incidence
+from .lap_joint import LapJoint
 from .l_butt import LButtJoint
 from .l_miter import LMiterJoint
+from .x_halflap import XHalfLapJoint
+from .t_halflap import THalfLapJoint
+from .l_halflap import LHalfLapJoint
 from .solver import ConnectionSolver
 from .solver import JointTopology
 from .solver import find_neighboring_beams
 from .t_butt import TButtJoint
-from .x_halflap import XHalfLapJoint
 
 __all__ = [
     "Joint",
     "beam_side_incidence",
+    "LapJoint",
     "BeamJoinningError",
     "TButtJoint",
     "LButtJoint",
     "LMiterJoint",
     "XHalfLapJoint",
+    "THalfLapJoint",
+    "LHalfLapJoint",
     "FrenchRidgeLapJoint",
     "JointTopology",
     "ConnectionSolver",