+++ src/compas_timber/ghpython/components/CT_Joint_Options_LButt/code.py
@@ -6,8 +6,12 @@ from compas_timber.ghpython import JointOptions
 
 
 class LButtJointOptions(component):
-    def RunScript(self, Gap, SmallBeamButts):
-        args = {"gap": Gap, "small_beam_butts": SmallBeamButts}
+    def RunScript(self, gap, small_beam_butts):
+        args = {}
+        if gap:
+            args["gap"] = gap
+        if small_beam_butts:
+            args["small_beam_butts"] = small_beam_butts
         options = JointOptions(LButtJoint, **args)
 
         return options