+++ src/compas_timber/ghpython/components/CT_Joint_Options_TButt/code.py
@@ -7,7 +7,9 @@ from compas_timber.ghpython import JointOptions
 
 class TButtJointOptions(component):
     def RunScript(self, Gap):
-        args = {"gap": Gap}
+        args = {}
+        if Gap:
+            args["gap"] = Gap
         options = JointOptions(TButtJoint, **args)
 
         return options