+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added new component `ShowFeatureErrors`.
 * Added new component `ShowJoiningErrors`.
 * Added `FeatureApplicator` classes which report errors during feature application.
+* added `L-HalfLapJoint`
+* added `T-HalfLapJoint`
+* added `ShowTopologyTypes` GH Component
 
 ### Changed
 
@@ -23,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+* removed `JointDef` GH components
+* removed `AutomaticJoint` GH Component. Joint rules are now input directly into `TimberAssembly`
 
 ## [0.4.0] 2024-01-24
 
@@ -47,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * added `BTLx` Grasshopper component
 * added `FrenchRidgeHalfLap` joint
 
+
 ### Changed
 
 * Changed `Beam` definition to include `blank_frame` and `blank_length` attributes 
@@ -61,7 +67,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-* Removed superfluous component `BeamFromCurveGuid`.
 
 ## [0.3.2] 2023-11-17